# [REVIEW] Improve lombscargle performance