… python/cusignal/spectral_analysis/_spectral.fatbin