The lombscargle kernels are instruction bound. This PR implements two changes to improve performance.

1. Create separate functions for `float` and `double`.
- This allows for specific type functions `cos` versus `cosf`
2. Exchange separate `sin` and `cos` with `sincos`

### Performance
I'm seeing roughly 1.35X on benchmarks (TITAN RTX)

Original
```bash
------------------------------------------------------------------------------------------- benchmark 'LombScargle': 8 tests -------------------------------------------------------------------------------------------
Name (time in ms, mem in bytes)                      Min               Max              Mean            StdDev            Median               IQR            Outliers       OPS            Rounds            Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_lombscargle_gpu[False-True-65536-1024]       0.0687 (1.86)     1.8579 (1.0)      1.5623 (1.0)      0.1977 (1.0)      1.5882 (1.0)      0.0053 (1.89)     253;2150  640.0871 (1.0)       14370           1
test_lombscargle_gpu[True-False-65536-1024]       0.0650 (1.76)     2.1015 (1.13)     1.7346 (1.11)     0.2515 (1.27)     1.7724 (1.12)     0.0038 (1.36)     337;2402  576.4964 (0.90)      15045           1
test_lombscargle_gpu[False-False-65536-1024]      0.0381 (1.03)     2.2263 (1.20)     1.8658 (1.19)     0.2597 (1.31)     1.9028 (1.20)     0.0028 (1.0)      509;2238  535.9557 (0.84)      25323           1
test_lombscargle_gpu[True-True-65536-1024]        0.0909 (2.46)     2.3116 (1.24)     1.9458 (1.25)     0.2501 (1.26)     1.9788 (1.25)     0.0055 (1.99)     198;2093  513.9325 (0.80)      10906           1
test_lombscargle_gpu[True-True-262144-1024]       0.0903 (2.44)     6.5473 (3.52)     5.9399 (3.80)     0.7739 (3.91)     6.0422 (3.80)     0.0082 (2.94)     188;1750  168.3522 (0.26)      10876           1
test_lombscargle_gpu[True-False-262144-1024]      0.0652 (1.76)     7.2030 (3.88)     6.6843 (4.28)     0.9836 (4.97)     6.8305 (4.30)     0.0071 (2.54)     324;3004  149.6036 (0.23)      14962           1
test_lombscargle_gpu[False-True-262144-1024]      0.0678 (1.83)     7.6394 (4.11)     7.1024 (4.55)     0.9095 (4.60)     7.2201 (4.55)     0.0091 (3.28)     238;2851  140.7972 (0.22)      14436           1
test_lombscargle_gpu[False-False-262144-1024]     0.0370 (1.0)      7.8454 (4.22)     7.3377 (4.70)     1.0137 (5.13)     7.4789 (4.71)     0.0035 (1.24)     493;4335  136.2823 (0.21)      26040           1
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

New
```bash
------------------------------------------------------------------------------------------- benchmark 'LombScargle': 8 tests -------------------------------------------------------------------------------------------
Name (time in ms, mem in bytes)                      Min               Max              Mean            StdDev            Median               IQR            Outliers       OPS            Rounds            Iterations
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_lombscargle_gpu[False-True-65536-1024]       0.0687 (1.82)     1.4664 (1.0)      1.1507 (1.0)      0.1460 (1.0)      1.1702 (1.0)      0.0039 (1.51)     255;2541  869.0109 (1.0)       14139           1
test_lombscargle_gpu[True-False-65536-1024]       0.0661 (1.76)     1.6832 (1.15)     1.2764 (1.11)     0.1841 (1.26)     1.3042 (1.11)     0.0058 (2.26)     344;1472  783.4522 (0.90)      14977           1
test_lombscargle_gpu[False-False-65536-1024]      0.0376 (1.0)      1.7259 (1.18)     1.3707 (1.19)     0.1903 (1.30)     1.3980 (1.19)     0.0026 (1.0)      511;2716  729.5492 (0.84)      25297           1
test_lombscargle_gpu[True-True-65536-1024]        0.0946 (2.51)     1.7857 (1.22)     1.4320 (1.24)     0.1877 (1.29)     1.4580 (1.25)     0.0067 (2.58)     202;1330  698.3253 (0.80)      10314           1
test_lombscargle_gpu[True-True-262144-1024]       0.0912 (2.42)     4.8279 (3.29)     4.3763 (3.80)     0.5703 (3.91)     4.4521 (3.80)     0.0067 (2.60)     189;1892  228.5028 (0.26)      10826           1
test_lombscargle_gpu[True-False-262144-1024]      0.0651 (1.73)     5.3860 (3.67)     4.9193 (4.27)     0.7209 (4.94)     5.0265 (4.30)     0.0069 (2.67)     325;2553  203.2797 (0.23)      15041           1
test_lombscargle_gpu[False-True-262144-1024]      0.0689 (1.83)     5.6726 (3.87)     5.2213 (4.54)     0.6744 (4.62)     5.3099 (4.54)     0.0114 (4.40)     239;3059  191.5244 (0.22)      14158           1
test_lombscargle_gpu[False-False-262144-1024]     0.0377 (1.00)     5.8595 (4.00)     5.3923 (4.69)     0.7445 (5.10)     5.4965 (4.70)     0.0054 (2.09)     494;4001  185.4481 (0.21)      26035           1
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```